+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Added
 - Converters for date, decimal and timestamp ORC types in Python and
   option to change them via Reader's and Writer's converters parameter.
 - Column object for accessing statistics about ORC columns.
+- An attribute to Reader for selected schema.
 
 Changed
 ~~~~~~~
@@ -16,6 +17,10 @@ Changed
 - Use timezone-aware datetime objects (in UTC) for ORC timestamps by default.
 - Wrapped C++ stripe object to Python Stripe.
 
+Fixed
+~~~~~
+
+- Decrementing reference for bytes object after reading from file stream.
 
 [0.1.0] - 2019-11-16
 --------------------